options-data/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-core==1.30.1
azure-functions==1.18.0
azure-storage-blob==12.19.1
openbb==4.1.6